Update README.md
Stegno_Block is a Python-based project that integrates steganography with blockchain technology to securely store and manage data. It uses image files to hide data (steganography) and a simple blockchain structure to ensure data integrity and immutability. The project includes a Flask web application for user interaction, user registration, and hash viewing functionalities.